Move retry_with_navigation and parse_response_json into zen_distill.py, delete mcp/utils.py and zen_actions.py
