+++ getgather/mcp/blindster.py
@@ -11,8 +11,7 @@ from getgather.mcp.dpage import (
     remote_zen_dpage_with_action,
 )
 from getgather.mcp.registry import GatherMCP
-from getgather.mcp.utils import retry_with_navigation
-from getgather.zen_actions import parse_response_json
+from getgather.zen_distill import parse_response_json, retry_with_navigation
 
 blindster_mcp = GatherMCP(brand_id="blindster", name="Blindster MCP")
