+++ getgather/mcp/garmin.py
@@ -7,8 +7,11 @@ from loguru import logger
 
 from getgather.mcp.dpage import remote_zen_dpage_with_action
 from getgather.mcp.registry import GatherMCP
-from getgather.zen_actions import parse_response_json
-from getgather.zen_distill import load_distillation_patterns, run_distillation_loop
+from getgather.zen_distill import (
+    load_distillation_patterns,
+    parse_response_json,
+    run_distillation_loop,
+)
 
 GARMIN_TIMEOUT_SECONDS = 15
