+++ getgather/mcp/tokopedia.py
@@ -11,9 +11,12 @@ from getgather.mcp.dpage import (
     remote_zen_dpage_with_action,
 )
 from getgather.mcp.registry import GatherMCP
-from getgather.mcp.utils import retry_with_navigation
-from getgather.zen_actions import parse_response_json
-from getgather.zen_distill import page_query_selector, zen_navigate_with_retry
+from getgather.zen_distill import (
+    page_query_selector,
+    parse_response_json,
+    retry_with_navigation,
+    zen_navigate_with_retry,
+)
 
 tokopedia_mcp = GatherMCP(brand_id="tokopedia", name="Tokopedia MCP")
